Adds the client as a singleton object, introduces, requests enum, request creator, and functionality for getting the current price for bitcoin. Still needs the price history request for bitcoin and test coverage

+++ Code/src/portfoliomanager/model/Crypto.java
@@ -1,5 +1,6 @@
 package portfoliomanager.model;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
 
@@ -15,7 +16,7 @@ public class Crypto {
 
 	private String name;
 	private Double currentPrice;
-	private HashMap<String, Double> historicalPrices;
+	private HashMap<String, BigDecimal> historicalPrices;
 	
 	/**
 	 * Instantiates a new Crypto object
@@ -32,7 +33,7 @@ public class Crypto {
 		}
 		this.name = name;
 		this.currentPrice = currentPrice;
-		this.historicalPrices = new HashMap<String, Double>();
+		this.historicalPrices = new HashMap<String, BigDecimal>();
 		
 	}
 	
@@ -87,7 +88,7 @@ public class Crypto {
 	 * 
 	 * @param historicalPrices the prices of the past
 	 */
-	public void setHistoricalPrices(HashMap<String, Double> historicalPrices) {
+	public void setHistoricalPrices(HashMap<String, BigDecimal> historicalPrices) {
 		this.historicalPrices = historicalPrices;
 	}
 	
@@ -97,7 +98,7 @@ public class Crypto {
 	 * @postcondition none
 	 * @return the historical prices of the cryoto
 	 */
-	public HashMap<String, Double> getHistoricalPrice() {
+	public HashMap<String, BigDecimal> getHistoricalPrice() {
 		return this.historicalPrices;
 	}
 	
@@ -112,10 +113,10 @@ public class Crypto {
 	 */
 	public HashMap<String, Double> getPriceForRange(int days) {
 		HashMap<String, Double> rangeHistorics = new LinkedHashMap<String, Double>();
-		for (HashMap.Entry<String, Double> entry : this.getHistoricalPrice().entrySet()) {
+		for (HashMap.Entry<String, BigDecimal> entry : this.getHistoricalPrice().entrySet()) {
 			if (days > 0) {
 				String date = entry.getKey();
-				Double price = entry.getValue();
+				Double price = entry.getValue().doubleValue();
 				rangeHistorics.put(date, price);
 				days--;
 			}
@@ -145,7 +146,7 @@ public class Crypto {
 		int prevDay = currentDay - 1;
 		String previousDay = String.valueOf(prevDay);
 		String yesterday = currentDate.replaceFirst("-[0-3][0-9]$", String.format("-%s", previousDay));
-		double yesterdaysPrice = this.historicalPrices.get(yesterday);
+		double yesterdaysPrice = this.historicalPrices.get(yesterday).doubleValue();
 		return ((this.currentPrice.doubleValue() - yesterdaysPrice) / yesterdaysPrice) * 100;
 	}
